device: get/print page protection details
This prints 4 words describing which pages are (not) write protected.
Each word corresponds to every 32 pages, and each bit in a word
corresponds to a specific page's protection (disabled) status. So if a
bit for a page is 0, then write protection is enabled for that page.

Example: 0xffffff80 0xffffffff 0xffffffff 0xffffffff
In the first word (which corresponds to the first 32 pages), page
protection is enabled only for the pages 0 to 6. There is no page
protection for the other pages.

Note: page size varies by model